feat(ir): §2.1 — emit_resolved walks ResolvedExpr directly; retire bridge
Closes the §2.1 from-scratch port. The runtime path is now
`ResolvedProgram → strata → emit_resolved → JIT` with no intermediate
ExprNode tree.

New module:
- `compiler/ir/emit_resolved.ts` (~580 LOC) — Emitter class walking the
  closed `ResolvedExprOpNode` union. Refs become operand kinds via
  decl-identity slot lookups (`inputRef → input slot`, `regRef →
  state_reg slot` or `array_reg`, `delayRef → state_reg slot=regCount+i`,
  `paramRef → param ptr` from the new `paramHandles` map). Param handles
  thread through `EmitSlots.paramHandles` (the §3-D1 gap noted in the
  original plan, now closed). Structural CSE keys ref-bearing nodes by
  decl identity rather than recursing through decl content (which would
  hit cycles via self-referential delay updates).

Reuses verbatim from emit_numeric: NInstr/NOperand/GroupInfo/FlatProgram
data types, BINARY_TAG/UNARY_TAG/CAST_RESULT/*_TAGS sets, type-inference
rules, structural-CSE intern table, array-loop emission patterns.

Deleted:
- compiler/emit_numeric.ts (759 LOC) — replaced.
- compiler/emit_numeric.test.ts (462 LOC) — coverage moves to the
  jit_interp_equiv, unified_ir, and stdlib integration tests.
- compiler/ir/lower_to_exprnode.ts (218 LOC) — `resolvedToSlotted` was
  the bridge that produced ExprNode trees for emit_numeric.
- compiler/ir/emit_node.ts (29 LOC) — `EmitExprNode` type alias for the
  bridge-time tree shape.

Wired:
- compile_resolved.ts now feeds `outputExprs: ResolvedExpr[]` and
  `registerExprs: (ResolvedExpr | null)[]` directly to
  `emitResolvedProgram`. No lowering step. Accepts an optional
  `paramHandles` ctx for the session path.
- program_type_builder.ts: input-default lowering inlined as
  `literalDefault` (handles literals + arithmetic + clamp/select; refs
  shouldn't appear in defaults so a ref-bearing op is a strata bug).
- flat_plan.ts, wasm_memory_layout.ts, emit_wasm.ts, emit_wasm.test.ts:
  type imports redirected from emit_numeric to emit_resolved.

Net delta: −1434 LOC.

Tests: 788 pass, 1 skip, 1 local-only fail (the same untracked
schema_audit). All 7 gateable subgraph tests pass under the new
emitter; LadderFilter (4 nested OnePoles + delay feedback)
audio-equivalent to interpret_resolved.
