+++ compiler/ir/compile_resolved.ts
@@ -5,32 +5,25 @@
  * instances, no combinators, no instance refs) and produces a
  * `tropical_plan_4` JSON consumed by the C++ JIT.
  *
- * Boundaries:
- *   - This function handles a *single* program. Session-level
- *     materialization (multiple instances + wiring + graph outputs) is
- *     `compileSession`'s job (D2; not yet implemented).
- *   - Post-strata invariant: `prog.body.decls` contains zero
- *     `instanceDecl` entries (`inlineInstances` lifts them all into a
- *     flat decl list of regs/delays). If an instanceDecl survives, the
- *     strata pipeline didn't run — caller bug.
- *
- * Implementation note: this iteration uses `resolvedToSlotted` (legacy
- * tree producer) + `emitNumericProgram` (the existing instruction
- * emitter) under the hood. A from-scratch ResolvedExpr-walking emitter
- * is still planned (PHASE_D_PLAN §2.1) but the dual-run gate
- * (emit.compat.test.ts) gives byte-equality across the boundary already
- * — so the rewrite is risk-free to land incrementally.
+ * Post-§2.1: walks `ResolvedExpr` directly via `emit_resolved.ts`. The
+ * legacy `resolvedToSlotted → emit_numeric` bridge is gone.
  */
 
-import type { ResolvedProgram, ResolvedExpr, OutputDecl, RegDecl, DelayDecl } from './nodes.js'
-import type { EmitExprNode as ExprNode } from './emit_node.js'
+import type { ResolvedProgram, ResolvedExpr, OutputDecl, RegDecl, DelayDecl, ParamDecl } from './nodes.js'
 import type { FlatPlan } from '../flat_plan'
-import { resolvedToSlotted } from './lower_to_exprnode.js'
 import { buildSlotMaps, type SlotMaps } from './slots.js'
-import { emitNumericProgram, type ScalarType } from '../emit_numeric.js'
+import { emitResolvedProgram, type EmitSlots, type ScalarType } from './emit_resolved.js'
+
+/** Optional per-emit context: param-handle bindings for FFI param/trigger
+ *  decls. compileSession populates this from the session's
+ *  paramRegistry; the per-program path (loadProgramAsType for stdlib
+ *  types, with no live session params) leaves it empty. */
+export interface CompileResolvedContext {
+  paramHandles?: Map<ParamDecl, { ptr: string }>
+}
 
 /** Compile a post-strata `ResolvedProgram` to `tropical_plan_4`. */
-export function compileResolved(prog: ResolvedProgram): FlatPlan {
+export function compileResolved(prog: ResolvedProgram, ctx: CompileResolvedContext = {}): FlatPlan {
   const slots = buildSlotMaps(prog)
 
   // Reject anything the strata pipeline should have removed. A surviving
@@ -42,14 +35,6 @@ export function compileResolved(prog: ResolvedProgram): FlatPlan {
     )
   }
 
-  const memo = new WeakMap<object, ExprNode>()
-  // Thread regCount through slots so resolvedToSlotted emits delayRefs
-  // directly as state-reg reads at the combined slot index, eliminating
-  // the previous separate `resolveDelayValues` post-pass.
-  const slotsWithRegCount = { ...slots, regCount: slots.regDecls.length }
-  const lower = (e: ResolvedExpr): ExprNode =>
-    resolvedToSlotted(e, slotsWithRegCount, memo)
-
   // ── Output expressions ──
   const outputExprByDecl = new Map<OutputDecl, ResolvedExpr>()
   for (const a of prog.body.assigns) {
@@ -58,12 +43,12 @@ export function compileResolved(prog: ResolvedProgram): FlatPlan {
       outputExprByDecl.set(a.target, a.expr)
     }
   }
-  const outputExprs: ExprNode[] = prog.ports.outputs.map(out => {
+  const outputExprs: ResolvedExpr[] = prog.ports.outputs.map(out => {
     const expr = outputExprByDecl.get(out)
     if (expr === undefined) {
       throw new Error(`compileResolved: program '${prog.name}' output '${out.name}' has no outputAssign.`)
     }
-    return lower(expr)
+    return expr
   })
 
   // ── Register update expressions ──
@@ -75,29 +60,30 @@ export function compileResolved(prog: ResolvedProgram): FlatPlan {
     if (a.target.op === 'delayDecl') delayUpdateByDecl.set(a.target, a.expr)
   }
 
-  const registerExprs: (ExprNode | null)[] = []
+  const registerExprs: (ResolvedExpr | null)[] = []
   const stateInit:     (number | boolean | number[])[] = []
   const registerNames: string[] = []
   const registerTypes: ScalarType[] = []
 
-  // Reg decls are state registers. Each contributes one slot; the update
-  // expression (if any) drives the per-tick writeback.
+  // Reg decls — each contributes one state-reg slot. The update expression
+  // (if any) drives the per-tick writeback.
   for (const d of slots.regDecls) {
     registerNames.push(d.name)
     registerTypes.push(regScalarType(d))
     stateInit.push(regInit(d))
     const u = regUpdateByDecl.get(d)
-    registerExprs.push(u === undefined ? null : lower(u))
+    registerExprs.push(u === undefined ? null : u)
   }
 
-  // Delay decls also occupy state-reg slots; their `update` lives on the
-  // decl (or a parallel nextUpdate may override).
+  // Delay decls — same state-reg layout (slots come AFTER regs in the
+  // unified state register sequence; emit_resolved threads `regCount`
+  // through `EmitSlots` so delayRefs resolve to `state_reg slot=regCount+delayIdx`).
   for (const d of slots.delayDecls) {
     registerNames.push(d.name)
     registerTypes.push(delayScalarType(d))
     stateInit.push(delayInit(d))
     const u = delayUpdateByDecl.get(d) ?? d.update
-    registerExprs.push(lower(u))
+    registerExprs.push(u)
   }
 
   // Input port types — typed input operands so int/bool inputs don't get floated.
@@ -105,13 +91,26 @@ export function compileResolved(prog: ResolvedProgram): FlatPlan {
     if (d.type === undefined) return 'float'
     if (d.type.kind === 'scalar') return d.type.scalar
     if (d.type.kind === 'alias')  return d.type.alias.base
-    // Array inputs: the element scalar drives operand typing for the
-    // element-wise ops; the array itself is bundled at a higher level.
     if (typeof d.type.element === 'string') return d.type.element
     return d.type.element.base
   })
 
-  const program = emitNumericProgram(outputExprs, registerExprs, stateInit, registerTypes, inputPortTypes)
+  const emitSlots: EmitSlots = {
+    inputs:       slots.inputs,
+    regs:         slots.regs,
+    delays:       slots.delays,
+    regCount:     slots.regDecls.length,
+    paramHandles: ctx.paramHandles ?? new Map(),
+  }
+
+  const program = emitResolvedProgram({
+    outputExprs,
+    registerExprs,
+    stateInit,
+    stateRegTypes: registerTypes,
+    inputPortTypes,
+    slots: emitSlots,
+  })
 
   // Compute array_slot_names — registers whose stateInit is a literal array.
   const arraySlotNames: string[] = []
@@ -154,9 +153,6 @@ function regInit(d: RegDecl): number | boolean | number[] {
   if (typeof init === 'number') return init
   if (typeof init === 'boolean') return init
   if (Array.isArray(init)) return init as number[]
-  // {op:'zeros', count: N} survives only when arrayLower didn't unroll
-  // (e.g. typeParam-shaped zeros — but those are rejected earlier). For
-  // a literal numeric count, expand to a zero-filled array.
   if (typeof init === 'object' && init !== null && (init as { op?: string }).op === 'zeros') {
     const count = (init as { count: ResolvedExpr }).count
     if (typeof count !== 'number') {
@@ -169,8 +165,7 @@ function regInit(d: RegDecl): number | boolean | number[] {
 
 function delayScalarType(_d: DelayDecl): ScalarType {
   // Delay decls today carry no scalar-type annotation in the resolved IR;
-  // the legacy emitter assumes float, which matches the C++ DAC contract
-  // (delays are always continuous float buffers in tropical_plan_4).
+  // the C++ DAC contract treats delays as continuous float buffers.
   return 'float'
 }
 
@@ -180,5 +175,5 @@ function delayInit(d: DelayDecl): number {
   return 0
 }
 
-void buildSlotMaps  // satisfies linter when SlotMaps is the only export consumed
+void buildSlotMaps
 export type { SlotMaps }
